protocol re-write to prep for video - each mux channel now has its own thread - SLIP is gone yay, the new way is much better - CPU % is about the same but the code is doing a lot more so that's good - still a bit rough; stuff will improve as work on video progresses - to implement next will be sender sending config to receiver

+++ include/endpoint.h
@@ -25,8 +25,8 @@ typedef struct {
   int lastPacketUTime;
 } endpoint_t;
 
-int endpoint_init (int (*onPacket)(const uint8_t*, int, int));
-int endpoint_send (const uint8_t *buf, int bufLen); // NOTE: this is not thread safe
+int endpoint_init (int (*onPacket)(const uint8_t*, size_t, int));
+int endpoint_send (const uint8_t *buf, size_t bufLen); // NOTE: this is not thread safe
 void endpoint_deinit (void);
 
 #endif
